prepare for hosting on vercel
